feat: Add TableSearch class. Issue #13
