+++ search-service/Cargo.toml
@@ -10,6 +10,8 @@ anyhow = { version = "1.0" }
 axum = { version = "0.6" }
 clap = { version = "4.3", features = ["derive"] }
 deadpool-postgres = "0.10.2"
+dict = "0.1.5"
+petgraph = "0.6.4"
 serde = { version = "1.0.188", features = ["derive"] }
 serde_json = "1.0.107"
 tokio = { version = "1", features = ["full"] }
