+++ search-service/src/relational/mod.rs
@@ -1,0 +1,2 @@
+pub mod general;
+pub mod tableSearch;
